Remove default loss function from generators
If the loss function is set as a default argument it will be loaded even
if not used, which is not desired.

FeatureLoss is used only for training, but was being loaded also for
inference, using a little more memory than required due VGG16 features.

This will not cause any harm because training and inference usage always
set gen_loss argument.